requirements.txt
fastapi
beautifulsoup4
uvicorn
cohere
openai
qdrant-client
langchain
pypdf
PyPDF2
unstructured
pydantic
python-multipart
python-dotenv
supertokens-python